Extracted curriculum-admin to topic-manager
