+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/access-pages-that-require-higher-privileges.spec.ts
@@ -38,19 +38,19 @@ test.describe('Logged-in Learner', function () {
 
   // The logged-in user cannot access the moderator page.
   test('should be restricted from accessing the moderator page', async function () {
-    await loggedInUser.navigateToModeratorPage();
+    await loggedInUser.navigateToModeratorPageAsLoggedInUser();
     await loggedInUser.expectErrorPage(401); // Expect a 401 Unauthorized error.
   });
 
   // The logged-in user cannot access the topics and skills dashboard.
   test('should be restricted from accessing the topics and skills dashboard', async function () {
-    await loggedInUser.navigateToTopicsAndSkillsDashboardPage();
+    await loggedInUser.navigateToTopicsAndSkillsDashboardPageAsLoggedInUser();
     await loggedInUser.expectErrorPage(401); // Expect a 401 Unauthorized error.
   });
 
   // The logged-in user cannot access the release coordinator page.
   test('should be restricted from accessing the release coordinator page', async function () {
-    await loggedInUser.navigateToReleaseCoordinatorPage();
+    await loggedInUser.navigateToReleaseCoordinatorPageAsLoggedInUser();
     await loggedInUser.expectErrorPage(404); // Expect a 404 Not Found error.
   });
