+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/interact-with-goals-in-learner-dashboard.spec.ts
@@ -128,7 +128,7 @@ test.describe('Logged-In Learner', function () {
   });
 
   test('should start and complete Chapter 1, then show updated progress (33%)', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.clickOnAddGoalsButtonInRedesignedLearnerDashboard();
@@ -153,7 +153,7 @@ test.describe('Logged-In Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
     await loggedInUser.expectGoalProgressToBeDisplayed('Place Values', 33);
 
@@ -163,7 +163,7 @@ test.describe('Logged-In Learner', function () {
   });
 
   test('should complete Chapter 2 and update progress to 67%', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
     await loggedInUser.clickOnGoalCard('Place Values');
 
@@ -177,7 +177,7 @@ test.describe('Logged-In Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.expectGoalProgressToBeDisplayed('Place Values', 67);
@@ -188,7 +188,7 @@ test.describe('Logged-In Learner', function () {
   });
 
   test('should complete final chapter and move goal to Completed section', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
     await loggedInUser.clickOnGoalCard('Place Values');
 
@@ -203,7 +203,7 @@ test.describe('Logged-In Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.expectGoalProgressToBeDisplayed('Place Values', 100);
@@ -220,7 +220,7 @@ test.describe('Logged-In Learner', function () {
   test('should display correctly on mobile viewport', async function () {
     await loggedInUser.setMobileViewport();
 
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.expectGoalCardToBeVisible('Place Values');
