+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/manage-goals-in-learner-dashboard.spec.ts
@@ -127,7 +127,7 @@ test.describe('Logged-In Learner - Manage Goals', function () {
   });
 
   test('should display empty Goals tab with title and Add Goals button', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.expectLearnerGreetingsToBe("loggedInUser1's Goals");
@@ -222,7 +222,7 @@ test.describe('Logged-In Learner - Manage Goals', function () {
   });
 
   test('should show goal card with 0% and Start button after adding goal', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInUser.navigateToGoalsSection();
 
     await loggedInUser.addGoalInRedesignedLearnerDashboard('Place Values');
@@ -252,7 +252,7 @@ test.describe('Logged-In Learner - Manage Goals', function () {
   });
 
   test('should highlight Goals tab in sidebar', async function () {
-    await loggedInUser.navigateToLearnerDashboard();
+    await loggedInUser.navigateToLearnerDashboardAsLoggedInUser();
 
     await loggedInUser.expectGoalsTabButtonToBeVisible();
     await loggedInUser.navigateToGoalsSection();
