+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/manage-in-progress-and-completed-lessons-and-skill.spec.ts
@@ -116,7 +116,7 @@ test.describe('Logged-in Learner', function () {
   });
 
   test('should display empty progress message when no lessons are in progress', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectSidebarTabToBeActiveAndContainButtonsInOrder(
       'Home'
     );
@@ -129,7 +129,7 @@ test.describe('Logged-in Learner', function () {
   });
 
   test('should select "Or Explore All Lessons in Classroom" button and navigate to /learn/math', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectClassroomButtonOnRedesignedLearnerDashboardToBePresent(
       true
@@ -146,7 +146,7 @@ test.describe('Logged-in Learner', function () {
     );
     await loggedInLearner.continueToNextCardAsLoggedOutUser();
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectScreenshotToMatch(
@@ -177,7 +177,7 @@ test.describe('Logged-in Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectScreenshotToMatch(
       'ProgressSectionInProgressWithOnlyChapter02'
@@ -185,7 +185,7 @@ test.describe('Logged-in Learner', function () {
   });
 
   test("should complete all the lessons of Place Value's Story and see Chapter 1 in the Completed Lessons section", async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectLessonCardProgressToBe(
@@ -205,7 +205,7 @@ test.describe('Logged-in Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectLessonCardProgressToBe(
       'Classroom Lessons',
@@ -224,7 +224,7 @@ test.describe('Logged-in Learner', function () {
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectScreenshotToMatch(
