+++ core/tests/playwright-acceptance-tests/specs/logged-out-learner/discover-the-website-and-navigate-to-math-classroom.spec.ts
@@ -67,7 +67,7 @@ test.describe('Logged-Out Learner', function () {
       'Algebra',
       'fractions'
     );
-    await curriculumAdmin.navigateToTopicAndSkillsDashboardPage();
+    await curriculumAdmin.navigateToTopicsAndSkillsDashboardPageAsTopicManager();
 
     await curriculumAdmin.openSkillEditor('fractions');
     await curriculumAdmin.navigateToSkillQuestionEditorTab();
