+++ core/tests/playwright-acceptance-tests/utilities/user/exploration-editor.ts
@@ -24,6 +24,7 @@ import {ExplorationEditorUtils} from '../common/exploration-editor-utils';
 import {RTEEditor} from '../common/rte-editor';
 import * as fs from 'fs';
 import * as path from 'path';
+import {StateEditorUtils} from '../common/state-editor-utils';
 
 const creatorDashboardPage = testConstants.URLs.CreatorDashboard;
 const baseUrl = testConstants.URLs.BaseURL;
@@ -114,18 +115,6 @@ const correctAnswerInTheGroupSelector = '.e2e-test-editor-correctness-toggle';
 const addNewResponseButton = 'button.e2e-test-add-new-response';
 const responseModalHeaderSelector = '.e2e-test-add-response-modal-header';
 const addAnotherResponseButton = 'button.e2e-test-add-another-response';
-
-const defaultFeedbackTab = 'a.e2e-test-default-response-tab';
-const openOutcomeFeedBackEditor = 'div.e2e-test-open-outcome-feedback-editor';
-const saveOutcomeFeedbackButton = 'button.e2e-test-save-outcome-feedback';
-const destinationSelectorDropdown = '.e2e-test-destination-selector-dropdown';
-const destinationWhenStuckSelectorDropdown =
-  '.e2e-test-destination-when-stuck-selector-dropdown';
-const saveDestinationButtonSelector = '.e2e-test-save-outcome-dest';
-const saveStuckDestinationButtonSelector = '.e2e-test-save-stuck-destination';
-const addDestinationStateWhenStuckInput = '.protractor-test-add-state-input';
-const outcomeDestWhenStuckSelector =
-  '.protractor-test-open-outcome-dest-if-stuck-editor';
 
 const mobileNavbarPane = '.oppia-exploration-editor-tabs-dropdown';
 const mobileTranslationTabButton = '.e2e-test-mobile-translation-tab';
@@ -787,9 +776,6 @@ export class ExplorationEditor extends BaseUser {
     await explorationEditor.dismissWelcomeModal(failIfMissing);
   }
 
-  // TODO(#22539): This function has a duplicate in exploration-editor.ts.
-  // To avoid unexpected behavior, ensure that any modifications here are also
-  // made in editDefaultResponseFeedbackInQuestionEditorPage() in question-submitter.ts.
   /**
    * Function to add feedback for default responses of a state interaction.
    * @param {string} defaultResponseFeedback - The feedback for the default responses.
@@ -801,36 +787,12 @@ export class ExplorationEditor extends BaseUser {
     directToCard?: string,
     directToCardWhenStuck?: string
   ): Promise<void> {
-    await this.expectElementToBeVisible(defaultFeedbackTab);
-    await this.clickOnElementWithSelector(defaultFeedbackTab);
-
-    if (defaultResponseFeedback) {
-      await this.updateDefaultResponseFeedbackInExplorationEditorPage(
-        defaultResponseFeedback
-      );
-    }
-
-    if (directToCard) {
-      await this.clickOnElementWithSelector(openOutcomeDestButton);
-      await this.select(destinationSelectorDropdown, directToCard);
-      await this.clickOnElementWithSelector(saveDestinationButtonSelector);
-      await this.expectElementToBeVisible(saveDestinationButtonSelector, false);
-    }
-
-    if (directToCardWhenStuck) {
-      await this.clickOnElementWithSelector(outcomeDestWhenStuckSelector);
-      // The '4: /' value is used to select the 'a new card called' option in the dropdown.
-      await this.select(destinationWhenStuckSelectorDropdown, '4: /');
-      await this.typeInInputField(
-        addDestinationStateWhenStuckInput,
-        directToCardWhenStuck
-      );
-      await this.clickOnElementWithSelector(saveStuckDestinationButtonSelector);
-      await this.expectElementToBeVisible(
-        saveStuckDestinationButtonSelector,
-        false
-      );
-    }
+    const stateEditorUtils = new StateEditorUtils(this);
+    await stateEditorUtils.editDefaultResponseFeedback(
+      defaultResponseFeedback,
+      directToCard,
+      directToCardWhenStuck
+    );
   }
 
   /**
@@ -1474,25 +1436,6 @@ export class ExplorationEditor extends BaseUser {
       default:
         throw new Error(`Unsupported interaction type: ${interactionType}`);
     }
-  }
-
-  /**
-   * Function to update the default response feedback for a state interaction.
-   * @param {string} defaultResponseFeedback - The feedback for the default responses.
-   */
-  async updateDefaultResponseFeedbackInExplorationEditorPage(
-    defaultResponseFeedback: string
-  ): Promise<void> {
-    await this.expectElementToBeVisible(openOutcomeFeedBackEditor);
-    await this.clickOnElementWithSelector(openOutcomeFeedBackEditor);
-    await this.clickOnElementWithSelector(stateContentInputField);
-    await this.typeInInputField(
-      stateContentInputField,
-      defaultResponseFeedback
-    );
-    await this.clickOnElementWithSelector(saveOutcomeFeedbackButton);
-
-    await this.expectElementToBeVisible(saveOutcomeDestButton, false);
   }
 
   /**
